+++ packages/main/all/src/Vitessce.tsx
@@ -24,7 +24,10 @@ export function Vitessce(props: any) {
     pluginJointFileTypes: pluginJointFileTypesProp,
   } = props;
 
-  const configUid = config?.uid;
+  // If config.uid exists, then use it for hook dependencies to detect changes
+  // (controlled component case). If not, then use the config object itself
+  // and assume the un-controlled component case.
+  const configKey = config?.uid || config;
   const configVersion = config?.version;
 
   const [configOrWarning, success] = useMemo(() => {
@@ -43,7 +46,7 @@ export function Vitessce(props: any) {
       ];
     }
   // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [configUid, configVersion]);
+  }, [configKey, configVersion]);
 
   const mergedPluginViewTypes = useMemo(() => ([
     ...baseViewTypes, ...(pluginViewTypesProp || []),