+++ packages/vit-s/src/VitessceGrid.js
@@ -38,6 +38,8 @@ const margin = 5;
  */
 export default function VitessceGrid(props) {
   const {
+    success,
+    configKey,
     rowHeight: initialRowHeight,
     config,
     theme,
@@ -75,28 +77,32 @@ export default function VitessceGrid(props) {
   }, [changeLayout, componentWidth]);
 
   // Update the view config and loaders in the global state.
+  // This effect is needed for the controlled component case in which
+  // the store has already been initialized with a view config,
+  // and we want to replace it with a new view config.
   useEffect(() => {
-    if (config) {
-      setViewConfig(config);
-      const loaders = createLoaders(
+    let newLoaders = null;
+    let newConfig = null;
+    if (success) {
+      newLoaders = createLoaders(
         config.datasets,
         config.description,
         fileTypes,
         coordinationTypes,
       );
-      setLoaders(loaders);
-    } else {
-      // No config found, so clear the loaders.
-      setLoaders({});
+      newConfig = config;
     }
-  }, [config, setViewConfig, setLoaders, fileTypes, coordinationTypes]);
+    setViewConfig(newConfig);
+    setLoaders(newLoaders);
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [success, configKey]);
 
   return (
     <div
       ref={containerRef}
       className={clsx(VITESSCE_CONTAINER, classes.vitessceContainer)}
     >
-      {layout && (
+      {layout ? (
         <VitessceGridLayout
           layout={layout}
           height={height}
@@ -112,7 +118,7 @@ export default function VitessceGrid(props) {
           onResize={onResize}
           onResizeStop={onResize}
         />
-      )}
+      ) : null}
     </div>
   );
 }